+++ TranSimCS/Model/Mesh.cs
@@ -12,7 +12,7 @@ namespace TranSimCS.Model {
 
     public class Mesh: IBVHElement{
         public List<VertexPositionColorTexture> Vertices { get; } = [];
-        public List<int> Indices { get; } = [];
+        public List<ushort> Indices { get; } = [];
         public IDictionary<int, object> Tags { get; } = new Dictionary<int, object>();
         public readonly MultiMesh? Parent;
 
@@ -32,22 +32,22 @@ namespace TranSimCS.Model {
             return distance < float.MaxValue;
         }
 
-        public Mesh(MultiMesh? parent = null, IEnumerable<VertexPositionColorTexture>? vertices = null, IEnumerable<int>? indices = null, IDictionary<int, object>? tags = null) {
+        public Mesh(MultiMesh? parent = null, IEnumerable<VertexPositionColorTexture>? vertices = null, IEnumerable<ushort>? indices = null, IDictionary<int, object>? tags = null) {
             this.Parent = parent;
             if(vertices != null) Vertices.AddRange(vertices);
             if(indices != null) Indices.AddRange(indices);
             if(tags != null) foreach(var row in tags) Tags.Add(row.Key, row.Value);
         }
-        public int AddVertex(VertexPositionColorTexture vertex) {
+        public ushort AddVertex(VertexPositionColorTexture vertex) {
             Vertices.Add(vertex);
             InvalidateAccelerationStructure();
-            return Vertices.Count - 1; // Return the index of the newly added vertex
+            return (ushort)(Vertices.Count - 1); // Return the index of the newly added vertex
         }
-        public void AddIndex(int index) {
+        public void AddIndex(ushort index) {
             Indices.Add(index);
             InvalidateAccelerationStructure();
         }
-        public void AddIndices(int[] indices) {
+        public void AddIndices(ushort[] indices) {
             Indices.AddRange(indices);
             InvalidateAccelerationStructure();
         }
@@ -57,7 +57,6 @@ namespace TranSimCS.Model {
             Vertices.AddRange(verts);
             InvalidateAccelerationStructure();
             return index;
-
         }
 
         public void Clear() {
@@ -66,7 +65,7 @@ namespace TranSimCS.Model {
             Tags.Clear();
             InvalidateAccelerationStructure();
         }
-        public void DrawTriangle(int a, int b, int c) {
+        public void DrawTriangle(ushort a, ushort b, ushort c) {
             AddIndex(a);
             AddIndex(b);
             AddIndex(c);
@@ -77,7 +76,7 @@ namespace TranSimCS.Model {
         /// </summary>
         /// <param name="vertices">List of vertices</param>
         /// <param name="indices">List of indices</param>
-        public void DrawModel(IList<VertexPositionColorTexture> vertices, IList<int> indices, IEnumerable<KeyValuePair<int, object>>? tags = null) {
+        public void DrawModel(IList<VertexPositionColorTexture> vertices, IList<ushort> indices, IEnumerable<KeyValuePair<int, object>>? tags = null) {
             ArgumentNullException.ThrowIfNull(vertices, nameof(vertices));
             ArgumentNullException.ThrowIfNull(indices, nameof(indices));
             int startVertexId = AddVerts(vertices.ToArray());
@@ -85,7 +84,7 @@ namespace TranSimCS.Model {
             int startingTriCount = startingIndex / 3;
             var indicesArray = indices.ToArray();
             for (int i = 0; i < indicesArray.Length; i++) {
-                indicesArray[i] += startVertexId;
+                indicesArray[i] += (ushort)startVertexId;
             }
             AddIndices(indicesArray);
             foreach (var kv in tags ?? []) {
